+++ src/main/kotlin/com/jfrog/conan/clion/conan/datamodels/Recipe.kt
@@ -1,0 +1,6 @@
+package com.jfrog.conan.clion.conan.datamodels
+
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class Recipe(val name: String, val versions: List<String> = listOf())