+++ src/main/kotlin/com/jfrog/conan/clion/icons/ConanIcons.kt
@@ -1,0 +1,8 @@
+package com.jfrog.conan.clion.icons
+
+import com.intellij.openapi.util.IconLoader
+
+object ConanIcons {
+    @JvmField
+    val Logo = IconLoader.getIcon("/icons/conan.svg", javaClass)
+}