+++ src/main/kotlin/com/jfrog/conan/clion/listeners/ConanCMakeSettingsListener.kt
@@ -1,0 +1,27 @@
+package com.jfrog.conan.clion.listeners
+
+import com.intellij.ide.util.PropertiesComponent
+import com.intellij.openapi.components.service
+import com.intellij.openapi.project.Project
+import com.jetbrains.cidr.cpp.cmake.CMakeSettings
+import com.jetbrains.cidr.cpp.cmake.CMakeSettingsListener
+import com.jfrog.conan.clion.cmake.CMake
+import com.jfrog.conan.clion.models.PersistentStorageKeys
+
+
+internal class ConanCMakeSettingsListener(private val project: Project) : CMakeSettingsListener {
+    override fun autoReloadChanged() {
+        println("autoReloadChanged")
+    }
+
+    override fun profilesChanged(old: List<CMakeSettings.Profile>, current: List<CMakeSettings.Profile>) {
+        val properties = project.service<PropertiesComponent>()
+        val autoAdd = (properties.getValue(PersistentStorageKeys.AUTOMATIC_ADD_CONAN, "false") == "true")
+        if (autoAdd) {
+            // get only the new profiles
+            current
+                .filter { new -> old.any { new.name == it.name }}
+                .forEach { CMake(project).injectDependencyProviderToProfile(it.name) }
+        }
+    }
+}