+++ src/main/kotlin/com/jfrog/conan/clion/models/LibraryData.kt
@@ -1,0 +1,8 @@
+package com.jfrog.conan.clion.models
+
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class LibraryData(
+        val libraries: HashMap<String, Library>
+)