+++ src/main/kotlin/com/jfrog/conan/clion/services/ConanPluginInit.kt
@@ -1,0 +1,12 @@
+package com.jfrog.conan.clion.services
+
+import com.intellij.openapi.components.service
+import com.intellij.openapi.project.Project
+import com.intellij.openapi.startup.StartupActivity
+
+
+class ConanPluginInit : StartupActivity {
+    override fun runActivity(project: Project) {
+        project.service<ConanService>().downloadLibraryData()
+    }
+}