+++ src/main/kotlin/com/jfrog/conan/clionplugin/conan/datamodels/Recipe.kt
@@ -1,6 +1,0 @@
-package com.jfrog.conan.clionplugin.conan.datamodels
-
-import kotlinx.serialization.Serializable
-
-@Serializable
-data class Recipe(val name: String, val versions: List<String> = listOf())