+++ src/main/kotlin/com/jfrog/conan/clionplugin/models/Component.kt
@@ -1,7 +1,0 @@
-package com.jfrog.conan.clionplugin.models
-import kotlinx.serialization.Serializable
-
-@Serializable
-data class Component(
-        val cmake_target_name: String? = null
-)