+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 <!-- Keep a Changelog guide -> https://keepachangelog.com -->
 
-# conan-clion-plugin Changelog
+# Conan CLion plugin changelog
 
 ## [Unreleased]
 ### Added
-- Initial scaffold created from [IntelliJ Platform Plugin Template](https://github.com/JetBrains/intellij-platform-plugin-template)
+- New Conan CLion plugin compatible with Conan 2.X